fix(query): use `replaceIn` to avoid mangling browser history
To ensure that the user retains a cleaner history of browsing
through Flat Viewer, set query parameters with `replaceIn` when:

repo-detail:
  - File/commit metadata is loaded

json-detail-container:
- A file is loaded and a tab chosen
- `onGridChange()` is invoked

The latter bears noting, as this means we will no longer retain
a user's history of manipulating the grid. Given that `onGridChange()`
is invoked when data is loaded into the grid, we think that ensuring
a cleaner overall history is more important to the user than retaining
the specific click history for grid manipuation, which the user may
find difficult to follow anyway.